Rate limit hat type off_t.

+++ relayd.h
@@ -684,7 +684,7 @@ struct protocol {
 	u_int8_t		 tcpflags;
 	int			 tcpbufsiz;
 	int			 tcpbacklog;
-	int			 tcprate;
+	off_t			 tcprate;
 	u_int8_t		 tcpipttl;
 	u_int8_t		 tcpipminttl;
 	u_int8_t		 tlsflags;
